move migration code folders, remove switch
