+++ pkg/descheduler/policyconfig_test.go
@@ -683,7 +683,7 @@ func TestV1alpha1ToV1alpha2(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
 			client := fakeclientset.NewSimpleClientset()
-			result, err := V1alpha1ToInternal(client, tc.policy, pluginregistry.PluginRegistry)
+			result, err := v1alpha1.V1alpha1ToInternal(client, tc.policy, pluginregistry.PluginRegistry)
 			if err != nil {
 				if err.Error() != tc.err.Error() {
 					t.Errorf("unexpected error: %s", err.Error())
